Simplify a few things.
Remove all types that are just slices of objects. They can be useful, but not as they are used right now, and then it clearly indicates that you get a slice back.

Main annoyance was that you were returning a pointer to a slice, which is unneeded (slices can be nil).

Return counts as int.